requirements.txt
# Core
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# AI
google-generativeai>=0.8.0

# Utilities
requests>=2.31.0